[#3] Do not xrefcheck CONTRIBUTING.md
Problem: CONTRIBUTING.md has a link to
https://github.com/tqtezos/baseDAO/issues/new
which causes xrefcheck to fail due to lack of permission.

Solution: a long-term solution should be on xrefcheck side.
We are going to implement it in near future, but until it's done
we add xrefcheck config that specifies that `.github/CONTRIBUTING.md`
should not be scanned.