[#32] Run network tests on CI properly
Problem: after `cleveland` update our tests can be run on network.
However, it requires some setup, for example, `tezos-client` must
be available, also we need to import a secret key with enough XTZ.

Solution: update `baseDAO-test` running on CI, so that it has all
necessary environment. For now we only run in local chain with
Delphi protocol. Carthage will likely be replaced in mainnet in 10
days. Running in Delphinet will be done separately.