[#29] Make storage more configurable from CLI
Problem: specification mentions that administrator address, voting
period and quorum threshold - are all configurable at contract
initialization, but in fact we have no such CLI options in our
executable, nor helpers in code that allow providing exactly those
options.

Solution:

1. Use `named` package in code to define storage constructor with
reasonable number of mandatory/optional arguments.

We leave administrator address as a mandatory parameter because
forgetting to provide it may cause the contract to belong to no one.
Voting period and quorum threshold are set as optional parameters,
because no one usually cares about them and they can be changed
afterwards.

2. Replace the existing `mkStorage` constructor that accepts balances
and operators - apparently it is not used, and it is suitable only for
testing purposes. If someone needs this behaviour, he can use field
selectors to override any fields.