[#29] Add a proposals counter to GameDAO
Problem: I have added contract extra to the GameDAO contract, however it
carries nothing. Since it is a sample contract, would be nice to add
some logic to it that would demonstrate the purpose of new field in
storage.

Solution: add a successful proposals counter, update the decision lambda
respectively.