[#56] Add permit to `vote` entrypoint
Problem: we want to implement something similar to TZIP-17 in `vote`
entrypoint, namely, caller should be able to create votes from someone
else's behalf, not only his own.

Solution: make `vote` accept additional `option (pair key signature)`.
If present, it designates signed vote parameter and public key for
signature verification.

Attach annotations in a special way so that it is possible to recognize
in a general case when an entrypoint has permit.

Modify existing `vote` entrypoint and the code it uses so that `sender`
is not used, rather the author of vote is carried explicitly; include
permit check in `vote`.