[#57] Reorganize `Parameter` to match LIGO version
Problem: LIGO forcibly performs fields reordering, and also supports
only specific layouts.
If we want to reuse existing tests, it would be simpler if at least
parameter type is the same in Lorentz and LIGO versions of the contract.

Solution: reorder constructors in `Parameter` type alphabetically like
in LIGO. Produce a left-balanced layout for `Parameter`.