[#86] Add custom entrypoints
Problem: we want to be able to include DAO-specific entrypoints to
contracts.

For instance, TreasuryDAO requires a way to transfer money to it
(which is recommended to be done via `%default unit` entrypoint), while
RegistryDAO should not accept money (there would be no way to withdrawal
them).

Solution: use `EpdDelegate`'s feature to include custom entrypoints, and
make parameter polymorphic over `otherParamater` type argument where
additional entrypoints can be specified.

Documenation for `EpdDelegate` does not recommend using it for supplying
default entrypoint in nested parts of parameter (justifying this by
modularity concerns), but nothing prevents us from including it, and
here we do this because this is exactly what we need.

Also we demonstrate how it works on `GameDAO` contract via adding
`Default` and reset entrypoints.