[#100] Implement management endpoints for LIGO Contract
Problem: Management entrypoints needs to be implemented. Right now there
are stubs left in `management.mligo` and they needs to be implemented.

Solution: Implement the entrypoints and add them to `management.mligo`.
A new file to hold the admin check and migration checks have been added.
As suggested, the migration check has been moved into this file. The
modules have been wrapped in ifdef directives to prevent the multiple
import issue.

For the `call_custom` entrypoint, we always expect the lambda to be of
type `(bytes * full_storage) -> return_with_full_storage`, and thus
delegates the decoding of the parameter to the entrypoint code. If the
corresponding entrypoint was not found in the entrypoint map, we throw
an 'ENTRYPOINT_NOT_FOUND' error. If the unpacking of the lambda fails,
we throw an 'UNPACKING_FAILED' error.