[#57] Change the default path used in `fetchContract`
Problem: on CI the default path in `fetchContract` seems to be
relative to the repo root. So the current one doesn't work and
can't work.
Solution: change it to a different path where the contract is already
put on CI.
It hopefully won't affect developers because we explicitly specify
`BASEDAO_LIGO_PATH` in `Makefile`.