[#116] LIGO RegistryDAO contract using BaseDAO
Problem: As an example, we would like to implement RegistryDAO contract
using ligo version of the BaseDAO contract.

Solution: Implement configuration lambdas for BaseDAO contract which
implements RegistryDAO behavior.

Originally, RegistryDAO contract requires a parameterized contract, but
since it is not clear how to do this in LIGO, the Key/Value types are
just included as a type alias in the source file.

Apart from this, the implementation of decision lambda is a bit hackish
where the two types of proposals are distinguished by looking at the
included keys in the proposal metadata, which is just a `(string, bytes)
map`.

Origination:
===========

During origination of BaseDAO, the configuration for RegistryDAO can be
loaded by making a storage with RegistryDAO lambdas included in
configuration and required map key in storage. A Michelson expression
for storage can be built during origination using the following LIGO
Command.

ligo compile-storage ligo/src/registryDAO.mligo base_DAO_contract
  'default_registry_DAO_full_storage(("tz1QozfhaUW4wLnohDo6yiBUmh7cPCSXE9Af" : address), ("tz1QozfhaUW4wLnohDo6yiBUmh7cPCSXE9Af" : address), 0n, 0n, 0n, 0n, 0n)'

Where the arguents to `default_registry_DAO_full_storage` function are
admin address, token_address and values of configuration parameters 'a',
'b', 's_max', 'c' and 'd'.