[#119] Add optional optimization to Makefile
Problem: Developers should have an easy way to run tests on the
optimized version of the contract.

Solution: Add a way to enable optimization to ligo/Makefile via
'OPTIMIZE' env variable, which should be set to 'true' to enable
optimization. Note that 'morley' binary is required for optimization.
Path to 'morley' binary can be specified via 'MORLEY' env variable.
It's expected to have 'morley' in PATH by default.