[#116] Include RegistryDAO storage in Makefile
Problem: We have a ligo-command to generate a BaseDAO storage with
lambdas implementing the RegistryDAO pattern. We would like to write the
output of this command to a file so that it can be used in the tests.

Solution: Add this command to the ligos Makefile and mention it in
README.