[#116] Add registry DAO tests
Problem: We should add some tests that checks the behavior of lambdas
implemented in LIGO, that implements registryDAO logic.

Solution: Add some infra, similar to the `readContract` function, to
read the file that contains the Michelson representation of storage of
registryDAO, that contains the lambds. Use it during the origination of
the BaseDAO contract in the tests. Implement test that checks
'CheckProposal' and `RejectedProposalReturnValue` lambdas works as
expected.