[#123] Add startup phase to LIGO baseDAO
Problem: the origination size of the contract is too big to fit the
limits. We could however modify the contract to store the entrypoint
code in its storage and load them in after the origination.

Solution: modify the contract and add a "startup mode" that can be
used right after origination to push entrypoints (custom or not)
inside the contract's storage.
The startup mode needs to be interrupted to move to the main contract
logic and can't be accessed again.