[#113] Add on-chain lookup-registry view to RegistryDAO LIGO version
Problem: We have to add an on-chain view to fetch values from registry.

Solution: Implement the entrypoint lambda in LIGO, and add it as a
custom entrypoint to registryDAO's default storage so that this
entrypoint is available upon origination. Right now the callback is
included as an `address` instead of a `contract` value, due to the
limitation of not being able to unpack `contract` types, and
we want to store this entrypoint as a packed lambda in the storage

Also adds a test to check the endpoint correctly looks up registry
values.