[#113] Add lookup-registry entrypoint to the Haskell version
Problem: We have added the look-up entrypoint to LIGO version. Now we
have to do the same with the Haskell version.

Solution: Implement the entrypoint and include it as part of the
`config` record while originating the contract. Here the view is
implemented as a `View` as we don't have to unpack the entrypoint from
storage. Also includes a test to check the behavior of the
lookup-entrypoint view and updates the registryDAO.md specification.