Use morley-large-originator to handle the large contract size
Problem: in #123 we had to deal with the excessive size of the Ligo
version of the contract, while the solution is working it's not
convenient to use, it also has excessive running times and may not
be enoung in the near future.
Moreover, the Lorentz version is currently just under the limit and
will need a solution as well.

Solution: we have a new tool, morley-large-originator that can
facilitate the origination of large contracts.
Moreover this integrated with our testing framework, so from this
moment we'll can make use of it.
Note: this reverts most of the changes made for #123, to not overlap
the two approaches.
On top of that, nettests have been temporarily deactivated on the
Lorentz version as well, due to the excessive running time.