[#67] Allow XTZ transfers while making a proposal
Problem: Some contrcts may want to ask for a certain fixed fee
in XTZ for submitting a proposal. Since `Propose` used to reject
such transactions, this logic was not possible.

Solution: Allow transferring XTZ while making a proposal.
The implementations are supposed to store the fee value in
`contract_extra` and use proposal checks to ensure `amount`
is greater than the expected fee.