[#138] Move `unfrozen_token_id` and `frozen_token_id` to storage
Problem: We should move the constants `{un,}frozen_token_id` to the
storage rather than have them inlined in the code.

Solution: Add these two parameters to the Ligo storage and update the
Ligo and Lorentz storages to contain s{Unf,F}rozenTokenId. In Ligo,
`grep` for the usage of these variables and properly use the ones coming
from storage. In Haskell (Ligo and Lorentz), grep for these variables
and update source files (but not tests) to use the parameters from
storage.