[#169] decouple Ligo and Lorentz version
Problem: the Haskell utilities (mainly the tests) for the Ligo
version of the contract use the Lorentz version as a dependency.
This allowed us to save a lot of duplication, however since the two
versions are no longer receiving the same updates maintaining this
relationship is becoming harder and will get worse.

Solution: decouple the two version by duplicating the Shared tests
as well as the reused types and utilities.
Note: this includes a direct copy of entire modules except for some
Lorentz-specific parts, but no other change other than the minimum
required ones have been done.