[#169] remove Lorentz ShareTest from its library
Problem: there are ShareTest modules in the Lorentz version library,
however these are no longer shared and should be in the tests only.

Solution: merge ShareTest modules' code inside the rest of the test
modules of the Lorentz version.
Additionally, a few definitions have been simplified and/or slightly
refactored for simplicity or to remove Shared/Ligo mentions.