[#170] Update tests
Problem: We changed the LIGO contract behavior during freeze/unfreeze
operation. We have to amend tests to take the new behavior into account.

Solution: Amend the tests, and account for the new behavior and cover
them in existing tests. We add a dummy FA2 contract that act like a
consumer contract, but only tracks calls to transfer entrypoint. The
state of this contract is checked to ensure that the BaseDAO contract
has made the expected FA2 calls.