[#170] Update build instructions to include governance_token
Problem: In the PR that implements BYO extension, we added Makefile
arguments `governance_token_address` and `governance_token_id`. But the
build instructions was not updated to reflect this. Some parts of the
documentation and specs are also bit outdated.

Solution: Update `docs/building.md` file to include the new arguments in
`make` command samples. In addition the default token address was
changed to an address of the form `KT*` instead of `tz*`, to better
reflect the fact that it is an address of a smart contract. Updated
specification document to amend outdated section and removed couple of
errors that no longer exist.