[#192] Change the `custom_entrypoints` type to `big_map`
Problem: The `custom_entrypoints` is currently defined as a `(string, bytes) map`.
This however poses a limitation, as `map` is strict and so the cost
in case of many key is too high.

Solution: Make `custom_entrypoints` as `big_map` instead of map, make
`DynamicRec` using `BigMap` instead, remove `DynamicRecBigMap`.