[#198] Add more assets to the auto prerelease step
Problem: In CI we have an auto prerelease step that creates a
prerelease based on master. Currently this only includes the
`baseDAO.tz` contract itself.

However, we can modify it to:
- compile all storage(s) with the default values
- generate the known metadata (for the ligo version only)

Solution: Add all trivialDAO, treasuryDAO, and registryDAO storages,
and add `known_metadata.json` to the pre-release step as assets.