[#203] Add token_id to stored operators
Problem: we need to be able to check that an operator for a token
holder is such for a specific token_id and not for any one available.
To do so we need to store the token_id for operators as well.

Solution: store the token_id for operators and use it to improve the
check.
Update the tests and TZIP-16 off-chain view.