Fix stages cycle to match the specification
Problem: there is a discrepancy between the specifiction and the
implementation (and tests) for the cycles: in the spec we say that
proposals stages are only odd and voting ones even, but the contract
behaves in the opposite way.

Solution: it seems to be more sensible to follow the specification,
although the current implementation is just as valid of an option, so
the contract code and tests are updated.
Additionally, tests have been commented more to show this more clearly.