[#230] Fix Makefile to use proper current timestamp
Problem: In the storage generation section of Makefile we use
'Tezos.now' to fetch the current timestamp. But it appears that the
'ligo compile' command is not able to fetch the current timestamp
properly.

Solution: Use the 'date' command from environment to fetch the unix
timestamp and use it as a timestamp during storage generation.