[#87] Update `OriginateFn` to use the record instead of tuple
Problem: Currently, there is a function called `OrignateFn` in the test that
is a type alias for a monad that return a tuple. Seems like tuple is not
the right type, since it requires a lot of modification when adding a
new value to the tuple.

Solution: Update `OriginateFn` to return a record `DaoOriginateData`
instead of tuple, update related tests.