[#207] Implement dynamic update of quorum-threshold
Problem: We should implement the dynamic update of the quorum threshold
based on the previous participation.

Solution: Introduce the concept of cycle, which is a proposing, voting
period pair. Then track the number of staked tokens in each cycle, and
use it to calculate the new value of quorum-threshold at the very first
propose call in each cycle. Since proposal can be flushed long after the
cycle in which it was proposed, we also store the quorum threshold for
the proposal cycle in the proposals. We use this stored quorum threshold
while evaluating the votes for this proposal.

Another main change in this PR is that we have stopped storing the
denominator for the fractions. This is to simplify the calculations,
mostly in update quorum procedure, and also to round off fractions to
the required precision.

Also updates documentation and specs.