[#250] Rename `rejected_proposal_return_value`
Problem: We currently have a `rejected_proposal_return_value` lambda to
decide what happens to tokens after a proposal is rejected.
However, the names suggest that the returned value of the function is the
amount of tokens to keep, whereas it's actually the amount to slash.

Solution: Rename `rejected_proposal_return_value` to
`rejected_proposal_slash_value`.