[#277] Replace max_votes with max_voters to limit the voters.
Problem: Right now we have a max_votes configuration to limit the number
of votes a proposal can have. But we would like to limit the number of
voters, since that is what adds to the size of the proposal. Also
right now we use a list to store voters in a proposal. But this
has a drawback that if a voter cast their votes multiple times, each of
them will add another item in the list.

Solution: Rename field `max_votes` to `max_voters` and change the voter
count check during the vote operation. Also Store the votes as a map
indexed on the voter address, and the vote type. This ensures that each
voter can only contribute to maximum of two items to the voters map.