[#277] Update tests
Problem: We have changed the max_votes value to max_voters to limit the
number of voters who can vote on a proposal. We have to update and add
new tests.

Solution: Add tests that check that the proposal correctly tracks the
voters with upvotes and downvotes. Also updates specs and Makefile.