[#258] Remove cleveland from library
Problem: Currently the `baseDAO-ligo-meta` `library` has a dependency on `cleveland`.
This is odd (as only the `test-suite` should have it) and the reason this exist is due
to the use of `Michelson.Test.Import`, which was the only way to read contract and storage
from the files produced by LIGO.
However since morley !565 MR has
been solved it's now possible to remove such dependency and only depend on the `morley` library instead.

Solution: Remove `cleveland` dependency from library, bump morley
version.