[#287] Add types to denote input/output for decisionLambda
Problem: Right now we don't have a type to denote the input/outputs for
the decision lambda. This was fine till now, because the lambda only
handled a couple of entities. But in this issue we might have to add a
yet another entity to the output of the lambda, namely the guardian.

Solution: Change the decision lambda to have a record type. Change the
corresponding Haskell types and default values.