[Chore] Fix bug with cabal validation
Problem: If you delete cabal file and then regenerate
it using generate-cabal-files.sh script the validation
will fail, because for new cabal files stack2cabal
adds hash line at the beginning of the file.

Solution: Add small change in the script to negate
this problem. Instead of moving a file before checking,
we now copy it.