[#284] Treat blocks as discrete entities
Problem: In our switch to using blocks to denote time interval, we had
to decide if we want to count blocks as discrete entities, or treat it
like we used to treat time. If we count the blocks as discrete entities,
a proposal that in originated in block `n` will be considered as being
alive for `1` block. But if we count blocks like we count time, then it
won't be considered as alive for `1` block until the next block. These
two options result in slightly different calculation when we want to
decide if a certain amount of blocks has elapsed.

Solution: We have decided that we want to treat blocks as discrete
entities. So we have included corresponding changes to the calculation
of time intervals.