[#263] Modify the makefile to generate variants
Problem: We have to implement code in Makefile to generate the
implementation module as per the variant requested via the build call.

Solution: Four new contract targets are added to the Makefile. This includes
the three BaseDAO variants, and a test contract variant that looks up
the implementation as lambdas stored in contract_extra. This is used to
export custom behavior for these lambdas to make some existing tests
work.

In addition to this, three new storage generation targets have also been
added to generate storage for the three variants. This still required
because the inital contract still differ for each variants for the extra
values.

For each variant, the appropriate source .mligo file is copied to the
'implementation.mligo' module, and the build command is called. This
results in the appropriate variant being generated. The same is done for
the contract as well as the storage targets.