[#263] Update Haskell code to work with LIGO changes
Problem: We have broken up the ligo contract into three different
contracts with slightly different parameter. We should make the Haskell
tests to work with these changes.

Solution: A type parameter, indicating the type of the custom entrypoint
parmeter for the variant was added to the main contract parameter. Types
representing the custom entrypoint argument was added to the types
module of all the variants.

Some tests used some predefined behavior of the lambdas. But now that
the lambdas were removed, we need a way to accomodate these tests. For
this a fourth variant, that fetched test lambdas from contract extra was
added. And in tests that checked for the core behavior, this variant was
used.

At this point the SMT tests are disabled.