[#306] Add a default entrypoint
Problem: We need a way to make it easy for people to send funds to the
contract.

Solution: Add a default entrypoint with a unit parameter so that funds
can be transferred to the contract without specifying an entrypoint or a
parameter. We also remove the custom entrypoint for treasury/registry
variants that previously served this purpose.