[#311] Remove voters limit and avoid returning tokens on flush
Problem: The cost of proposals handling increases linearly with the
number of voters due to the fact that we unfreeze voters token on flush.

As a result, we want to to change this behavior to remove the voters
limit and avoid returning tokens on flush.

Solution: Remove returning voter tokens, remove voter limit, and add
`Unfreeze_staked` entrypoints.