[#316] Remove proposal limit and use a linked list for sorting
Problem: There is currently a limit on the number of concurrent ongoing
proposals, due to the fact that proposals need to be sorted (for `flush`)
and that this is done using a `set`, which cannot grow indefinitely without
causing gas cost problems.

Solution: Remove `proposal_key_list_sort_by_level` and use a doubly linked
list instead to store the proposals in order. Update the tests and SMT
accordingly.