[#329] Implement LambdaDAO
Problem: We need to implement the LambdaDAO as per the spec.

Solution: Implement LambdaDAO as per the spec. But we had to slightly
deviate from the spec at the following places.

1. Proposal handler input aka `ph_input` needed to include
`proposal_info` field of type `propose_params` so that the handler can
compute proposal_key. This was required in the registryDAO
implementation using LambdaDAO to properly implement the
'registry_affected' procedure to compute the proposal_key of the
proposal that affected a certain key.

2. Proposal handler output aka `ph_output` needed to include a guardian
field of type `address option` to allow the updation of contract
guardian from within the handlers.