[Chore] Update 'haskell.nix' to use the upstream from 'master'
Problem: the nix sources are not taking 'haskell.nix' from the
'master' branch of the 'input-output-hk' repo, but rather from
a branch of our fork with a (very old) fix.
Updating this source with 'niv' causes issues in CI.

Solution: switch 'haskell.nix' to use the 'master' branch of
the 'input-output-hk' repo.
Additionally update it and 'hackage.nix' to the latest
available versions.